address review comments about test case locations
